fix tools
